server: re-factored code & organised endpoints

+++ app/server/src/controller/authCtrl/resetPasword.ts
@@ -1,8 +1,8 @@
-import { Request, Response, NextFunction } from "express";
-import path from "path";
-import { retriveToken } from "../../lib/createToken";
-import { asyncHandler, ErrorResponse } from "../../lib/utils";
-import User from "../../models/User";
+import { Request, Response, NextFunction } from 'express';
+import path from 'path';
+import { retriveToken } from '../../lib/createToken';
+import { asyncHandler, ErrorResponse } from '../../lib/utils';
+import User from '../../models/User';
 
 // @desc      Reset Pasword
 // @route     PUT /api/v1/auth/resetpassword/:resettoken
@@ -10,28 +10,27 @@ import User from "../../models/User";
 
 export const resetPasword = asyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
-    const { token } = await retriveToken(
-      "resetpassword",
-      req.params.resettoken
-    );
+    const token = await retriveToken(req.params.resettoken);
+    if (token.type != 'resetPassword')
+      return next(ErrorResponse(400, 'Resource not found'));
 
     const user = await User.findById(token.userId);
 
     if (!user) {
-      return next(ErrorResponse(400, "Resource not found"));
+      return next(ErrorResponse(400, 'Resource not found'));
     }
     user.password = req.body.newPassword;
     await user.save();
     await token.delete();
     return res.json({
-      status: "ok",
-      message: "Password changed successfully",
+      status: 'ok',
+      message: 'Password changed successfully',
     });
   }
 );
 
 export const resetPaswordPage = asyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
-    res.sendFile(path.join(__dirname, "ResetPassword.html"));
+    res.sendFile(path.join(__dirname, 'ResetPassword.html'));
   }
 );
